+++ src/trubbel/manifest.json
@@ -10,7 +10,7 @@
 	"description": "Just some random things.",
 	"author": "Trubbel",
 	"maintainer": "Trubbel",
-	"version": "2.4.1",
+	"version": "2.5.7",
 	"search_terms": "trubbel",
 	"website": "https://twitch.tv/trubbel",
 	"created": "2025-01-06T23:29:54.496Z"
