+++ src/trubbel/settings/remove-things.js
@@ -238,7 +238,7 @@ export class RemoveThings extends FrankerFaceZ.utilities.module.Module {
     if (this.settings.get("addon.trubbel.remove-things.stories")) {
       this.style.set("hide-stories-left-nav-expanded", "#side-nav [class*=\"storiesLeftNavSection--\"] { display: none !important; }");
       this.style.set("hide-stories-left-nav-collapsed", "#side-nav :is([style]) :has([class*=\"storiesLeftNavSectionCollapsedButton--\"]) { display: none !important; }");
-      this.style.set("hide-stories-following-page", "div[class^=\"Layout-sc-\"] > [data-simplebar=\"init\"] > div:nth-child(3) > :has(h2[class*=\"sr-only\"]:is([class^=\"CoreText-sc-\"])) { display: none !important; }");
+      this.style.set("hide-stories-following-page", "div[class^=\"Layout-sc-\"]:has(> [data-simplebar=\"init\"] > .simplebar-scroll-content > .simplebar-content[style] h2.sr-only) { display: none !important; }");
     } else {
       this.style.delete("hide-stories-left-nav-expanded");
       this.style.delete("hide-stories-left-nav-collapsed");
