+++ src/trubbel/utils/constants/selectors.js
@@ -1,3 +1,4 @@
+export const FOLLOWING_CHANNELS_SELECTOR = "[data-a-target=\"following-channels-tab\"] [class*=\"ScTitle-sc-\"]";
 export const MOST_RECENT_CARD_SELECTOR = ".offline-recommendations-video-card:has([class*=\"offline-recommendations-video-card-border-\"]) :is(.tw-aspect)";
 export const MOST_RECENT_SELECTOR = ".offline-recommendations-video-card";
 export const TIMESTAMP_SELECTOR = ".timestamp-metadata__bar + p";
