+++ src/trubbel/utils/graphql/follows_totalCount.gql
@@ -1,0 +1,8 @@
+query Trubbel_GetTotalChannelFollowCount {
+	currentUser {
+		id
+		follows {
+			totalCount
+		}
+	}
+}
